feat: takein requestSchema to docs component
